pyproject.toml
[project]
name = "05-shippment-info"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "bs4>=0.0.2",
    "flask>=3.1.2",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-groq>=1.0.1",
    "langgraph>=1.0.3",
    "playwright>=1.56.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
